summarize/requirements.txt
black==23.3.0
langchain==0.0.152
openai==0.27.5
python-dotenv==1.0.0
tiktoken==0.3.3